+++ lua/github-theme/extra/init.lua
@@ -1,5 +1,7 @@
 package.path = "./lua/?/init.lua;./lua/?.lua"
 
+local types = require("github-theme.types")
+
 ---Write content to file.
 ---@param str string content of file.
 ---@param file_name string path of file.
@@ -18,11 +20,11 @@ local extras = {
   tmux = "tmux"
 }
 
-for _, style in ipairs({"dark", "dimmed", "light", "dark_default", "light_default"}) do
+for _, style in pairs(types.gt.ThemeStyle) do
   local config = {theme_style = style, transform_colors = true}
   for extra, ext in pairs(extras) do
-    local plugin = require("github-theme.extra." .. extra)
+    local theme = require("github-theme.extra." .. extra)
     local file_name = "extras/" .. extra .. "/" .. style .. "." .. ext
-    write(plugin[extra](config), file_name)
+    write(theme(config), file_name)
   end
 end