+++ lua/github-theme/extra/kitty.lua
@@ -1,13 +1,9 @@
 local util = require("github-theme.util")
-local config_module = require("github-theme.config")
-
-local M = {}
 
 ---Generate github theme for kitty terminal.
----@param config github-theme.Config
-function M.kitty(config)
-  config = config or config_module.config
-  local colors = require("github-theme.colors").setup(config)
+---@param cfg gt.ConfigSchema
+return function(cfg)
+  local colors = require("github-theme.colors")(cfg)
 
   local kitty = util.template([[
 # github colors for Kitty
@@ -55,5 +51,3 @@ color17 ${bright_red}
 ]], colors)
   return kitty
 end
-
-return M