+++ lua/github-theme/init.lua
@@ -1,12 +1,13 @@
 local util = require("github-theme.util")
-local theme = require("github-theme.theme")
-local config_module = require("github-theme.config")
+local config = require("github-theme.config")
 
+---@param user_config gt.ConfigSchema
 local function setup(user_config)
-  if user_config then config_module.apply_configuration(user_config) end
+  if user_config then config.apply_configuration(user_config) end
 
   -- Load colorscheme
-  util.load(theme.setup(config_module.config))
+  local theme = require("github-theme.theme")(config.schema)
+  util.load(theme)
 end
 
 return {setup = setup}